requirements.txt
fastapi==0.103.0
nats-py==2.3.1
pydantic==2.3.0
uvicorn==0.23.2
